+++ includes/globals.php
@@ -5,7 +5,7 @@
  */
 
 // Current CMS version
-const VERSION = '1.6.0';
+const VERSION = '1.6.1';
 
 /**
  * Display copyright on the admin dashboard.